Movie dataset
The Movie Database (TMDb) is a community-built movie and TV database. 

In this presentation, based on this dataset various questions will be answered for the curious minds. For example, what is the distribution of year movies were produced? who were the top actors, directors, and movie genres? what are the distributions of movies runtime, average vote?